removed meta name="viewport"
No mobile version of this site is present, and including this
meta-element only causes strange behaviors in mobile devices. for
example the http://www.mozilla.org/media/img/sandstone/bg-sand.png does
not repeat itself more the 320px on a iPhone. and on other devices when
page loads it is zoomed in to a not user-friendly level.

And it's safe to remove.