Fix typo. There a element with form-details id not a class.
Apparently this was working anyway because we include
footer-email-form.js which does -among other things- the same
thing. Probably though this is referring to another form somewhere in
the site, so we better have the code also in this contribute-form.js
to make sure that this won't break if code in footer-email-form.js
gets modified in the future.