Remove logic to do anything without a valid token
If the user passes no token to /existing, or the token is not valid,
earlier versions of this code would still allow them to sign up for
new newsletters. That's a change from the current behavior. We've
decided instead of that, we're going to just let them request a link
in the email or unsubscribe all. But that won't happen until a later
bug, so for now, just go back to the current behavior in this case.

Bug 845904