Bug 835506: Add facebookapps.utils unit tests
Tests for app_data_query_string_encode handle normal conversion of a dict into an app_data query string for Facebook apps.

Tests for unwrap_signed_request function handle empty signed request, changing the locale from Facebook format (underscore) to Playdoh format (hyphen) and a normal unwrap of a signed request object. Two utility functions are also added for use by the tests, one that creates a dummy dictionary to encode and the other to encode it.

Tests for get_best_locale function handle supported locales and a variety of cases for unsupported locales: unsupported locales of a language with support and completely unsupported languages (returns default locale). Tests also assert that in all the tested cases that a supported locale is always returned and that the active locale is not changed (the best locale is merely returned). To pass tests, locales are now always converted to lowercase before comparison. Also corrected the exception type when language isn't found in supported locales.

Tests for js_redirect verify that the returned response has the appropriate HTML elements for JavaScript redirection with the new assert_js_redirect assertion method.