Bug 850753: Port /products/download.html to Bedrock
This is an INCOMPLETE start on porting the download transitional/thank-you page to Bedrock. The basic page is copied over, including much of the ugly old javascript.
The images for the mac and windows download instructions have not yet been added, and some other design tweaks will be required.
The download page also relies on a 'lang' URL variable that gets eaten by funfactory. For now, the page uses a 'dlang' var and you have to change it manually to actually get the page to work.
All of the JS should be heavily reviewed and probably refactored