Always show subscribed newsletters on /existing
Short version: on /existing, don't worry about what languages a
newsletter supports if the user is already subscribed to it.
Always show it, and don't complain if the form language is changed
to one that the newsletter doesn't support, the subscribed box
is left checked, and the form is submitted.  Let the user continue
to subscribe to it.

Also made some fixes to the javascript that hides and displays the
newsletters on the page. It wasn't really working at all.

Bug 857271