Cope with long language codes from ET
Our internal list of newsletter languages is all 2-letter codes, like
'es'. But some of our users have language preferences set in ET to
longer codes, like 'es-ES'. This was not getting matched up and the
result was when they loaded the /existing page, the initial language
selected was whatever was first in the list, like Indonesian.