Fix contribute email tests.
These tests were relying on files in "locale"
to either exist or not, and these change often.
I've added mocks that removes their dependence
on these sortof external files.

This does include an upgrade of the Mock library
to 1.0.1. You'll need to `pip install mock==1.0.1`
in order to get your local tests to pass.