Bug 893144 - /newsletter/existing/ styling
The styling sheet on /newsletter/existing ws messed up - the
descriptions were all bolded and the newsletter titles were small.

Reverted styling of the TH part of the table to what it was in an
earlier version of newsletter.less. No way to tell what happened
to it due to rebasing on the original branch.