Bug 893165 - /id/newsletter formatting
The issue was that in the header 'Read all about it in our
<span>newsletter</span>', the <span> and </span> each were causing a
line break. In languages like Indonesian where the word order was
changed, e.g.:

    Baca selengkapnya di <span>nawala</span> kami

this was breaking the header up into three short lines.

It turned out that header spans were being styled in newsletter.less
as "display: block". Getting rid of that solves the problem.