Remove duplicate messages from /newsletter/existing
The template for /existing displayed the Django messages so we could
tell the user when their token was bad or other errors occurred.

In bug 899837, we changed the base template to display Django messages,
resulting in the /existing page showing the messages twice.

Remove the display of messages from the /existing page's template so
we only see the messages once, via the base template.