Fix bug 924347: Fx download button default to en-US.
Previously for release Fx it would search for the build
in the locale that had the largest version number that
wasn't a pre-release or ESR. For some locales this was
returning a very old release. This alters this to only
return the latest release, and thus causes the button
to always give the latest release by defaulting to en-US
if no localized build is available.