Bug 975476: Update redirects for python-nss docs.
We now have a github repo setup using github-pages
to publish the docs. The developer (@jdennis) has
permissions on it, and they will be updated there.

I've made the redirect for the docs more specific since
after examining the archive it's clear that this directory
structure is all that was in there, so there wouldn't be
any valid links to any URLs between "python-nss" and "html".

I made the trailing slash on the MDN redirect optional since
that form of the URL is the one used on the PyPI page for the
package and was redirecting to a 404 at the archive.