Hide the info bar on the SUMO article.
SUMO recently turned on Tabzilla's update nag, and turned ours off.
The bug for that is Bug 1092335. Now, if a user clicks on the "Update"
button in Tabzilla, they get sent to the SUMO article, which proceeds to
nag them again. This patch hides the update bar on the target url.