Bug 1088752: Add Shape of the Web app to serve JSON.
Shape of the Web is a standalone site that pulls data from JSON files.
In order to make the l10n easier initially, we’re going to host the
JSON files on Bedrock to take advantage of the existing l10n process.

Also adds a middleware to handle adding CORS headers to the locale
redirects when loading the JSON.