Add a script to fix bug 1117866
This script checks every icalendar file present in <bedrock>/media/caldata
and validates it. Currently, it would run every time a commit is done on master,
but that can be changed to just run on commits to .ics files.

There's also a failure in MexicanHolidays.ics, that I have left as-is for the author
to fix - this would be a good test to check/verify that all folks use the same file encoding.

Moved back check_calendars.py to scripts dir

Fix Mexican calendar

Ideally, this should be done by caldata maintainers. We are fixing this
now that we go live and avoid build failures preventing merge into
master.