Fix bug 1155810: Remove funfactory.
Well... actually it mostly just moves most of funfactory
from an external lib to bedrock.base. But at least this way
we can remove the bits we don't need and be ready to make
the changes we need to upgrade to Django 1.8.

I can't come close to writing as perfect a commit message
for this as @willkg did for fjord, so I'll just leave this here.

https://github.com/mozilla/fjord/commit/9d7c97b