Deploy to webheads in series.
This is intended to cause commander to run `deploy_app`
for each host in series instead of parallel like it does now.
It also adds a 5s wait before allowing the next code push to happen.

TODO: Add a check for a 200 response code from the host before proceeding to
the next host.