Fixed the bug #1200464
Changed the redirection setting which navigates users with ja-locale to
wrong Japanese localized releasenotes.

Updated the redirection rule which navigate ja-locale users to Japanese localized releasenotes, to navigate the accesses to Fennec's releasenotes to their Japanese version as well

update redirection rules with review comments

shortened the redirection rule which redirects Japanese releasenotes to mozilla.jp. Also, a bug in a redirection rule is fixed

fixed a redirection bug, which doubled slash in the URL when it redirect ja-locale users to beta releasenotes