Send no email for success@example.com addresses
Partnership form should not send the email if the submitted
address is at success@example.com. This is a reserved testing domain
and will never have valid email accounts. This is specifically
so that the selenium tests for this page will not send mail
to the list for these tests.