correct minor grammar nit ("when if")
Under the Diversity and Inclusion section, in the sublist for "Some Mozillians may identify with activities or organizations that do not support the same inclusion and diversity standards as Mozilla. When this is the case:", there's an occurrence of "when if" in the sentence "when if (a) and (b) are met, other Mozillians should treat this as a private matter, not a Mozilla issue."

I suspect this was meant to be either "when" or "if" but not both, since either makes sense on its own, while both are ungrammatical.