Fix supervisor running in dev
This was done before using a different process type in
Procfile, but only "web" and "cmd" types are served as
HTTP traffic, so that wouldn't work. This new direction
should work for any instance for which we'd like to use
supervisord.

Also reunified the cron.py file and used arguments to
determine which jobs are scheduled.