Remove spurious whitespace in links from 2015 annual report
When the closing element of a link is on the next line, the whitespace
is rendered as part of the link.  This is especially visible when a
link is followed by something other than whitespace, as here with the
the Webmaker link, which is rendered as "Webmaker ."  Avoid this
problem by moving the </a> to the same line.