Downgrade Requests to fix LinkChecker
Turns out LinkChecker checks for a good version of python-requests
in a dumb way, so it sees version "2.13.0" as being lower than
"2.2.0". Also the project seems to have been abandoned, so there
isn't a newer release, though it seems this issue has been fixed
in the master branch.

Easiest fix is to just downgrade requests to the latest version
that is < 2.10.0, which is what I've done.