Fix #5928: Run full assets build in PR tests
Circle can't create symlinks during a Docker build
for some reason >:( which means that we can't do a full
production build on Circle. The assets build step has now
moved into a step that is run in Circle so that an error
there will fail the build and also removes a step in the
build process.