Fix #6036: Handle no pocket article image in template
The fallback image is currently passed through `lazy_img`
but could be used in other contexts, so this should be handled
at the view or template layer.

Also increase the number of articles fetched to 8 and limit the
display to 4. This will help if one of the fetched articles fails
to store for some reason.