Fix #6036: Handle no pocket article image in template (#6039)
* Fix #6036: Handle no pocket article image in template

The fallback image is currently passed through `lazy_img`
but could be used in other contexts, so this should be handled
at the view or template layer.

Also increase the number of articles fetched to 8 and limit the
display to 4. This will help if one of the fetched articles fails
to store for some reason.

* Fix warnings from storing non-aware datetime objects for pocket

* Only update pocket articles when timestamp changes

This will prevent the DB from appearing to change on every
run of `update_pocketfeed`.