[#6444] Add Firefox Concert Series landing page (#6546)
* [#6444] Add Firefox Concert Series landing page

A landing page for the Firefox Concert Series, promoting Firefox Accounts. This uses our new FxA OAuth flow so you can create an account and sign in from any browser. It's not tied to Sync or UITour so you'll see a "create account" form even if you're signed into Firefox.

It's English only so no l10n, and the concert stream will only be available in the US anyway. Uses geolocation to determine if the user in the US and can access the stream; all non-US countries get an apology. If JS fails or geolookup fails we show the concert info by default (the idea is to assume the user has access unless we can determine otherwise; at showtime it's livenation's problem).