Allow normal "img/" path prefix for helper functions
Functions like `l10n_img`, `platform_img`, `lazy_img`, and
`high_res_img` all wanted the image path to be passed in without
the `img/` prefix, but the normal `static` function did want that.
This change allows the developer to be consistent and use the same
path for all of the image calls.