Remove SECRET_KEY from all .env files
It turns out we don't need them because we have a default value
in settings/base.py. This covers the cases where we need to run
management commands during the build and test phases. Real
SECRET_KEY values are set in all of our running instances of
the site.

Fix #7792