Catch URL errors in the Firefox Nightly Notes Feed
An invalid version string in release notes was causing a
NoReverseMatch error to be thrown in the Nightly Notes Feed
(/firefox/nightly/notes/feed/). This was also being indexed
for the sitemap, so it caused the build to fail. This both
skips releases with this error in the feed and removes the
feed from the sitemap.