A couple of debugging messages for bedrock release notes process
We've had a few instances in dev of the db not being updated with release notes.
This change makes it a bit easier to read the logs, by putting a success in the run-upload-db.py file, and adding newlines.
Non-quiet release notes jobs will output the current hashes, and new hash.