Fix docker image tagging
Raphael discovered that the sitemap was behind. I found that this
is because the `prod-latest` and git-tag-based docker tags weren't
being pushed to Docker Hub. This is because we moved to promoting
images instead of building fresh for stage and prod. Because of this
the docker hub push was being skipped for stage and prod. This change
ensures that the tagging and pushing happens regardless.