Fix #10434: Add country_code to the context for all pages
Allows for a query param to override the country (e.g. `?geo=de`).

Also removes the old Maxmind stuff since we're unlikely to use that
again, and add a GeoTemplateView class to make it easy to switch
template based on request country. Adds docs for all of this as well.