NOTICKET: Making local doc building work again following urllib upgrade
In order to `make livedocs`, one needs to install requirements/docs.txt locally. However, upgrading urllib3 from 1.25.8 to 1.26.5 in #10555 broke the installation of `requests`.

This changeset upgrades `requests`, adds a hashed dependency introduced by requests, and also upgrades attrs, so that installation works on Python 3.9. Python 3.10 is not yet supported for this installation - it was about getting it back on its feet first.