10690: Fix leak of MPL comment into newsletter management form
This changeset stops the newsletter preferences form at /newsletter/existing/<hash>/ from showing a reference to MPL2.0, which was leaking from a template partial being used to render the form.

Two templates have been updated, one of which was clearly leaking. The other was updated for consistency and belt-and-braces safety

The key issue was that the template had a Jinja2-style comment, which was not being skipped because the standard Django form renderer was being used.

Switching to a jinja2 renderer did not work, so instead the fix has been to swap the comment style to be Django Template Language style AND also add a magic string so that the pre-commit hook does not think that the templates in question lack the licence text.