10593: Add support for an optional frontend-specific Sentry DSN
In some situations/environments we want to separate out backend and frontend sentry alerts, to make spotting issues easier.

This changeset adds support for a frontend-specific DSN, which will default to be the same DSN as the one for the backend (and the single one currently in use).

We can add config for specific environments only (eg Production) and leave all other envs using the same DSN for FE and BE