10650 + 10504: Contentful-powered VPN Resource Centre pages (#10697)
* Code VPN Knowledge center (#10504)

* Add GA to voting widgit

* Formatting change

* 10650: Main pass at backend changes to consume a new Compose-based pagePageResourceCentre entity

* The data from the Contentful API client for a Compose entry is now JSON for the nested page, which requires slightly different handling
* Rich-text response rendered as a list of HTML-containing entities
* Separate SEO model data is merged into the info dict, so the overall internal client response doesn't change a great deal
* slug field already existed on ContentfulEntry model - now it is populated

Note that this changeset attempted to NOT break any of the existing homepage syncing behaviour, and so has retained an allow-list of page types to sync. However, because one of those page types is simply `page` we now get connected pages (eg pageHome, connected from connectHomepage) being added to the DB. We're headed towards syncing all pages, so it will be poss to drop the allow-listing, which will follow in a separate piece

* 10650: Stop rendering each chunk of the VRC page's rich-text as a separate entry

* 10650: Initial wiring up of he FE work done by Nathan for the VRC

Works, using data from Contentful, but needs some tweaks to remove hard-coded vals from the FE mocks (eg VPN pricing etc)

* VRC landing page: only show the second split if there was something after the first split

* 10650: Rename ContenfulEntry.language to .locale

* 10650: Add support for previewing pagePageResourceCenter from Contentful

* 10650: Improve error handling and locale mismatches

Use active_locales to hook in automatic redirection behaviour if other locales are selected.
Also update what we do if we really can't find the right article based on the slug - in that situation we should 404, rather than error out

* 10650: Minor tweaks following initial/draft PR review

* Change rule to h3 over class - contentful fix

* Add vpn pricing helper to VRC #10504

* 10650: Support capturing the Category and Tags for an Entry, as well as a 'Classification' field

... which for the VPN Resource Centre is populated with the value of the Project field.
(The idea of ContentfulEntry.classification is that it lets us group our resource centre
pages by project/topic, with cagtegories and tags being a lower level of targettings)

* 10650: Improve error handling and reporting in update_contentful management command

* Raise a verbose exception at an early stage if there is no useful linked Entry available
* Add 'errors' as an explicit metric for the result of the update_contentful command, to help with debugging

* 10650: Update VRC landing page with category filtering

* Show all categories in pick-list, if there is more than one category specified
* Filter by categories
* Update manager's get_entries_by_type method to return ContentfulEntry records, not dicts of data
* Add in support for filtering entries by 'classification' field, so that we can re-use the same page type in Contentful for multiple areas within the site, rather than having to have product-specific page types

Also drop the redundant and unlikely-to-be-needed get_page method on ContenfulEntryManager

Also ensure related articles are in the same locale

* 10650: Refactor ContentfulPreviewView test to work in CI

Hot reloding the urlconf with different settings wasn't working in when the tests were run for CI, so this approach instead deliberately makes available the dev-only urls that are included when settings.DEV is True. It's not perfect, but it works reliably

While testing, it became apparent that relying on the SEO object for the page title is short-sighted if there is no override title set in the SEO object.

So, better to try to get it from the page object itself, with some fallbacks for existing page support and for preview titles (existing behaviour)

Co-authored-by: Nathan Barrett <30009669+nathan-barrett@users.noreply.github.com>