10744: Avoid duplicates when generating list of VRC categories based on available data
While we're not yet using/revealing any categories in the VRC, we have code that provides a list of them for filtering. This code was not avoiding duplicates, so this changeset remedies that, via TDD.

There were different ways to approach this, but the separate tracking set of 'seen' categories seemed the simplest.