10814: Add support for pinging a new Dead Man's Snitch (DMS) after DB update
* Change script mode to allow continue execution even if one of the management 
commands has a non-zero exit value
* Set a sentinel to true if we encounter a problem during one of those 
management commands
* If the sentinel is false AND we have a specific, new, DMS url 
configured, ping it to confirm the update script went without problem. 
NOT pinging it will trigger an alert within MEAO circles.

Note that the new DMS snitch URL will come from the environment, but is 
not loaded in as a Django setting.