10886: Update detection of environment to use settings.APP_NAME, rather than settings.PROD
Also clarifies that settings.PROD is meant as an indication that the code is running
in production mode (which can/does happen on the Dev, Stage and Prod deployments).
This contrasts with `settings.DEV = True`, used locally or on a demo deployment to
enable developer-helping behaviour that we don't want in production.

Also updates some comments around the GO_ACTIONS for the Contentful sync.